修复刷新没有加载用户信息数据的问题，清除缓存后重载加提示

+++ src/store/modules/modules/account.js
@@ -30,10 +30,10 @@ export default {
       commit('get_snsapi_userinfo')
       commit('get_userinfo')
 
-      if (state.userinfo && state.userinfo['openid']) return
+      if ((state.userinfo && state.userinfo['openid']) || !payload) return
 
       window.localStorage.clear()
-      
+
       let snsapi_userinfo = await service['cdcHttpClient.getAccessToken'](payload)
       let user_info = await service['cdcHttpClient.getUserinfo']({
         params: {
